added retry

+++ web/helpers/sqlite_database.py
@@ -286,13 +286,13 @@ def get_temperature2():
             for items in group:
                 for item in items:
                     if item[0] == 10:
-                        _sum_hum_air += int(item[2])
-                        _sum_temp_air += int(item[1])
+                        _sum_hum_air += float(item[2])
+                        _sum_temp_air += float(item[1])
                     if item[0] == 11:
-                        _sum_temp_out_air += int(item[1])
-                _avr_temp_air = round(_sum_temp_air / _group_len, 1)
-                _avr_hum_air = round(_sum_hum_air / _group_len, 1)
-                _avr_temp_out_air = round(_sum_temp_out_air / _group_len, 1)
+                        _sum_temp_out_air += float(item[1])
+            _avr_temp_air = round(_sum_temp_air / _group_len, 1)
+            _avr_hum_air = round(_sum_hum_air / _group_len, 1)
+            _avr_temp_out_air = round(_sum_temp_out_air / _group_len, 1)
 
             grouped_by_line_id[key] = {'temp_air': _avr_temp_air,
             'hum_air': _avr_hum_air,
